+++ docs/CHANGELOG.md
@@ -32,6 +32,7 @@
 - Ticket #213 : lors de l'édition des informations d'un profil/d'une configuration, l'appui sur Entrée valide les modifications
 - Ticket #26 : ajout de la possibilité de désactiver la création de profils/configurations de secours à la suppression/l'import
 - Ticket #212 : si l'installeur est lancé alors que LifeCompanion est déjà lancé, l'installation est bloquée
+- Ticket #54 : si LifeCompanion est lancé une nouvelle fois avec une configuration à importer, alors l'import se déclenche dans l'application déjà ouverte
 
 ## 1.4.5 - 21/02/2023